fix: #4 - `docker-compose.yaml`の変更
