+++ docker-compose.yaml
@@ -22,14 +22,36 @@ services:
     container_name: backend_container
     # to debug
     ports:
-      - "1323":"1323"
+      - "1323:1323"
     networks: 
       app_net:
         ipv4_address: 192.168.0.3
     environment:
       TZ: "Asia/Tokyo"
       URI_FRONTEND: "http://192.168.0.2:8080"
-    
+      MYSQL_USER: trip
+      MYSQL_PASSWORD: trip_app
+      MYSQL_HOST: 192.168.0,4
+      MYSQL_PORT: 3306
+      MYSQL_DATABASE: trip_db
+
+  db:
+    image: mysql
+    ports:
+      - 3306:3306
+    networks: 
+      app_net:
+        ipv4_address: 192.168.0.4
+    command: mysqld --character-set-server=utf8 --collation-server=utf8_unicode_ci
+    volumes:
+      - ./backend/db/init:/docker-entrypoint-initdb.d
+    environment:
+      MYSQL_ROOT_PASSWORD: triptrip
+      MYSQL_DATABASE: trip_db
+      MYSQL_USER: trip
+      MYSQL_PASSWORD: trip_app
+
+
 networks: 
   app_net:
     driver: bridge
